Add missing charm requirement: requests
This charm was depending on requests but did not install it in its venv.
Due to the previous default for `include_system_packages` of `true`, it
was previously relying on the system-wide version. That default changed
recently so it's now necessary to include it in the wheelhouse. It's
also safer to do so.