templates/config.toml: render registry mirror endpoints
According to doc[0], for custom registries, we need to add mirror entry
and endpoint, which is missing in current template.

This patch also introduces a new config option `host`, which allows user
to provide registry host addr or name.

The key is named `host` because it's normally supposed to be registry
host address, e.g.: `my.registry.com` or `10.10.10.10:5005`.
When endpoint is absent, containerd will use this value to try
endpoints.
When endpoint is present, this key can be a simple name, e.g.: `my_registry`.

[0]: https://github.com/containerd/cri/blob/master/docs/registry.md#configure-registry-endpoint

Signed-off-by: Joe Guo <joe.guo@canonical.com>